Спринт № 11, учёл замечания ревьюера
