+++ app/src/main/java/com/example/playlistmaker/RetrofitClient.kt
@@ -1,0 +1,16 @@
+package com.example.playlistmaker
+
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+
+class RetrofitClient() {
+    private val iTunesBaseUrl = "https://itunes.apple.com"
+    private val retrofit = Retrofit.Builder()
+        .baseUrl(iTunesBaseUrl)
+        .addConverterFactory(GsonConverterFactory.create())
+        .build()
+    private val iTunesService = retrofit.create(ItunesApi::class.java)
+    fun getITunesService(): ItunesApi {
+        return iTunesService
+    }
+}
