+++ app/src/main/res/values-night/themes.xml
@@ -3,10 +3,10 @@
     <style name="Base.Theme.PlaylistMaker" parent="Theme.MaterialComponents.DayNight.NoActionBar">
         <!-- Customize your dark theme here. -->
         <!-- <item name="colorPrimary">@color/my_dark_primary</item> -->
-        <item name="colorPrimary">@color/black</item>
+        <item name="colorPrimary">@color/dark_gray</item>
         <item name="colorOnPrimary">@color/white</item>
         <item name="colorPrimaryVariant">@color/white</item>
-        <item name="colorSecondary">@color/black</item>
+        <item name="colorSecondary">@color/dark_gray</item>
         <item name="colorOnSecondary">@color/white</item>
         <item name="colorOnTertiary">@color/dark_gray</item>
         <item name="colorTertiary">@color/dark_gray</item>
